% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3AreaChart_dollar.R
\name{C3AreaChart_dollar-shiny}
\alias{C3AreaChart_dollar-shiny}
\alias{C3AreaChart_dollarOutput}
\alias{renderC3AreaChart_dollar}
\title{Shiny bindings for C3AreaChart_dollar}
\usage{
C3AreaChart_dollarOutput(outputId, width = "100\%", height = "400px")

renderC3AreaChart_dollar(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a C3AreaChart_dollar}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using C3AreaChart_dollar within Shiny
applications and interactive Rmd documents.
}
