% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3AreaChart_dollar.R
\name{C3AreaChart_dollar}
\alias{C3AreaChart_dollar}
\title{<Add Title>}
\usage{
C3AreaChart_dollar(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
