% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3AreaChart_perc.R
\name{C3AreaChart_perc}
\alias{C3AreaChart_perc}
\title{<Add Title>}
\usage{
C3AreaChart_perc(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
