% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3PieChartabsolute.R
\name{C3PieChartabsolute}
\alias{C3PieChartabsolute}
\title{<Add Title>}
\usage{
C3PieChartabsolute(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
