% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3rotatedArticleBarChart.R
\name{C3rotatedArticleBarChart}
\alias{C3rotatedArticleBarChart}
\title{<Add Title>}
\usage{
C3rotatedArticleBarChart(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
