% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3rotatedCustomerBarChart.R
\name{C3rotatedCustomerBarChart}
\alias{C3rotatedCustomerBarChart}
\title{<Add Title>}
\usage{
C3rotatedCustomerBarChart(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
