% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3simpleAreaChart.R
\name{C3simpleAreaChart}
\alias{C3simpleAreaChart}
\title{<Add Title>}
\usage{
C3simpleAreaChart(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
