% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3simpleCategoricalBarChart.R
\name{C3simpleCategoricalBarChart}
\alias{C3simpleCategoricalBarChart}
\title{<Add Title>}
\usage{
C3simpleCategoricalBarChart(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
