% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3simpleKWBarChart.R
\name{C3simpleKWBarChart}
\alias{C3simpleKWBarChart}
\title{<Add Title>}
\usage{
C3simpleKWBarChart(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
