% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3simpleMonthBarChart.R
\name{C3simpleMonthBarChart}
\alias{C3simpleMonthBarChart}
\title{<Add Title>}
\usage{
C3simpleMonthBarChart(
  dataset,
  colors,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
