% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileDependency.R
\name{fileDependency}
\alias{fileDependency}
\title{Create a file dependency that can be accessed by the client using javascript.}
\usage{
fileDependency(filename, version = "0.0.1")
}
\description{
Create a file dependency that can be accessed by the client using javascript.
}

