% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDependency}
\alias{getDependency}
\title{Deprecated: use [widgetDependencies()] instead.}
\usage{
getDependency(name, package = name)
}
\arguments{
\item{name}{Widget name.}

\item{package}{Package name.}
}
\description{
Deprecated: use [widgetDependencies()] instead.
}
\keyword{internal}
