% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_bower_pkg}
\alias{install_bower_pkg}
\title{Install bower package to inst/htmlwidgets/lib}
\usage{
install_bower_pkg(pkg)
}
\arguments{
\item{pkg}{name of bower package to install}
}
\description{
This function uses bower to install a javascript package along with
its dependencies.
}

