% Generated by roxygen2 (4.0.1): do not edit by hand
\name{installBowerPkg}
\alias{installBowerPkg}
\title{Install bower package to inst/htmlwidgets/lib}
\usage{
installBowerPkg(pkg)
}
\arguments{
\item{pkg}{name of bower package to install}
}
\description{
This function uses bower to install a javascript package along with
its dependencies.
}

