% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JS}
\alias{JS}
\title{Mark character strings as literal JavaScript code}
\usage{
JS(...)
}
\arguments{
\item{...}{character vectors as the JavaScript source code (all arguments
will be pasted into one character string)}
}
\description{
This function \code{JS()} marks character vectors with a special class, so
that it will be treated as literal JavaScript code when evaluated on the
client-side.
}
\examples{
library(htmlwidgets)
JS('1 + 1')
list(x = JS('function(foo) {return foo;}'), y = 1:10)
JS('function(x) {', 'return x + 1;', '}')
}
\author{
Yihui Xie
}

