% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeShinyOutput}
\alias{makeShinyOutput}
\title{Create a shiny output function for a widget}
\usage{
makeShinyOutput(name, package = name, defaultWidth = "100\%",
  defaultHeight = "400px")
}
\description{
Create a shiny output function for a widget
}

