% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{prependContent}
\alias{prependContent}
\alias{appendContent}
\title{Prepend/append extra HTML content to a widget}
\usage{
prependContent(x, ...)

appendContent(x, ...)
}
\arguments{
\item{x}{An HTML Widget object}

\item{...}{Valid \link[htmltools]{tags}, text, and/or
\code{\link[htmltools]{HTML}}, or lists thereof.}
}
\value{
A modified HTML Widget object.
}
\description{
Use these functions to attach extra HTML content (primarily JavaScript and/or
CSS styles) to a widget, for rendering in standalone mode (i.e. printing at
the R console) or in a knitr document. These functions are NOT supported when
running in a Shiny widget rendering function, and will result in a warning if
used in that context. Multiple calls are allowed, and later calls do not undo
the effects of previous calls.
}
