% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{projectTemplate}
\alias{projectTemplate}
\title{Creates an RStudio Project Template}
\usage{
projectTemplate(path, ...)
}
\arguments{
\item{path}{Destination path where this project will be created.}

\item{...}{Additional parameters, currently not used.}
}
\description{
Creates an RStudio Project Template
}
