% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolveSizing}
\alias{resolveSizing}
\title{Resolve widget sizing policy}
\usage{
resolveSizing(x, sp, standalone, knitrOptions = NULL)
}
\arguments{
\item{x}{The widget object whose size is to be determined. It may have $width
and $height directly on it, which means we should obey those.}

\item{sp}{The sizing policy to use.}

\item{standalone}{Logical value indicating whether the widget is being
rendered in a standalone context (where it's the only thing on the page;
this is usually via `print.htmlwidget()`).}

\item{knitrOptions}{Object representing the knitr options passed to us via
`knit_print`. If we're not doing a `knit_print` right now, then the value
should be `NULL`.}
}
\value{
A list that is guaranteed to have `width` and `height` values, each of
  which is either a number or CSS unit string. If `standalone=TRUE` then the
  list will also have a `runtime` value that is a list, that contains two
  nested lists `viewer` and `browser`. Each of those in turn has `width`,
  `height`, `padding` (between 1 and 4 numbers), and `fill` (`TRUE`/`FALSE`).
}
\description{
Take a widget object and sizing policy, and some other contextual details,
and figure out what width/height to use, if possible. Some decisions may need
to be deferred until runtime; include any metadata that's needed for that
decision in the result as well.
}
\examples{
x <- list(
  sizingPolicy = list(
    defaultWidth = 800,
    defaultHeight = 500,
    padding = 15,
    viewer = list(
      fill = TRUE,
      padding = 0
    ),
    browser = list(
      fill = FALSE,
      defaultWidth = 960,
      defaultHeight = 600,
      padding = 20
    ),
    knitr = list(
      # Actually knitr$defaultWidth and knitr$defaultHeight
      # are ignored if figure = TRUE
      defaultWidth = 800,
      defaultHeight = 600,
      figure = TRUE
    )
  )
)

# Sizing for standalone mode
resolveSizing(x, x$sizingPolicy, TRUE, NULL)
# Sizing for knitr
resolveSizing(x, x$sizingPolicy, FALSE,
  list(out.width.px = 150, out.height.px = 100))

# Explicit width/height provided by user--overrides any
# default width/height
x$width <- 300
x$height <- 250
resolveSizing(x, x$sizingPolicy, FALSE,
  list(out.width.px = 150, out.height.px = 100))
}
\keyword{internal}

