% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplewidget.R
\name{scaffoldSimpleWidget}
\alias{scaffoldSimpleWidget}
\title{Scaffold simple widgets}
\usage{
scaffoldSimpleWidget(name, bowerPkg = NULL, edit = interactive(), dir_)
}
\description{
Scaffold simple widgets
}

