% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffoldWidget}
\alias{scaffoldWidget}
\title{Create implementation scaffolding for an HTML widget}
\usage{
scaffoldWidget(name, bowerPkg = NULL, edit = interactive())
}
\arguments{
\item{name}{Name of widget}

\item{bowerPkg}{Optional name of \href{https://bower.io/}{Bower} package upon
which this widget is based. If you specify this parameter then bower will
be used to automatically download the widget's source code and dependencies
and add them to the widget's YAML.}

\item{edit}{Automatically open the widget's JavaScript source file after
creating the scaffolding.}
}
\description{
Add the minimal code required to implement an HTML widget to an R package.
}
\note{
This function must be executed from the root directory of the package
  you wish to add the widget to.
}
