% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetid.R
\name{setWidgetIdSeed}
\alias{setWidgetIdSeed}
\title{Set the random seed for widget element ids}
\usage{
setWidgetIdSeed(seed, kind = NULL, normal.kind = NULL)
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}

\item{kind}{character or \code{NULL}.  If \code{kind} is a character
    string, set \R's RNG to the kind desired.  Use \code{"default"} to
    return to the \R default.  See \sQuote{Details} for the
    interpretation of \code{NULL}.}

\item{normal.kind}{character string or \code{NULL}.  If it is a character
    string, set the method of Normal generation.  Use \code{"default"}
    to return to the \R default.  \code{NULL} makes no change.}
}
\description{
Set a random seed for generating widget element ids. Calling this
function rather than relying on the default behavior ensures
stable widget ids across sessions.
}
