% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sizingPolicy}
\alias{sizingPolicy}
\title{Widget sizing policy}
\usage{
sizingPolicy(defaultWidth = NULL, defaultHeight = NULL, padding = NULL,
  viewer.defaultWidth = NULL, viewer.defaultHeight = NULL,
  viewer.padding = NULL, viewer.fill = TRUE, viewer.suppress = FALSE,
  viewer.paneHeight = NULL, browser.defaultWidth = NULL,
  browser.defaultHeight = NULL, browser.padding = NULL,
  browser.fill = FALSE, knitr.defaultWidth = NULL,
  knitr.defaultHeight = NULL, knitr.figure = TRUE)
}
\description{
Widget sizing policy
}

