% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_json}
\alias{to_json}
\title{Convert data frames to JSON more flexibly}
\usage{
to_json(df, orient = "columns", json = TRUE)
}
\description{
The toJSON function in RJSONIO converts a dataframe to an object of
arrays by column. Sometimes, we need an array of objects. This function
provides more flexible options to convert a dataframe to JSON. It also
allows one to return an R object which when passed to the toJSON function
will return the desired JSON structure.
}
\examples{
d <- data.frame(x = c(1, 2), y = c("a", "b"))
to_json(d)                     # {"x":[1,2],"y":["a","b"]}
to_json(d, orient = 'records') # [{"x":1,"y":"a"},{"x":2,"y":"b"}]
to_json(d, orient = 'values')  # [[1,"a"],[2,"b"]]
}

