% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{widgetDependencies}
\alias{widgetDependencies}
\alias{widgetDependencies.character}
\alias{widgetDependencies.htmlwidget}
\title{Get HTML dependencies for a widget}
\usage{
widgetDependencies(x, ...)

\method{widgetDependencies}{character}(x, ..., package = x)

\method{widgetDependencies}{htmlwidget}(x, ...)
}
\arguments{
\item{x}{A widget name (character) or an \code{htmlwidget} object.}

\item{...}{Currently unused.}

\item{package}{Name of the package containing the widget (only used when
\code{x} is a character string). Defaults to the widget name.}
}
\value{
A list of \code{\link[htmltools]{htmlDependency}} objects.
}
\description{
Get HTML dependencies for a widget
}
