% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{widgetList}
\alias{as.widgetList}
\alias{widgetList}
\title{Widget lists}
\usage{
widgetList(...)

as.widgetList(x)
}
\arguments{
\item{...}{HTML widget objects created by \code{createWidget()} or widget
packages, and (optionally) HTML elements created from \pkg{htmltools}}

\item{x}{a list to be coerced to a widget list}
}
\description{
Combine multiple HTML widgets into a list, and all widgets will be rendered
on the same page when the widget list is printed.
}

