% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{as.iframe}
\alias{as.iframe}
\title{Render widget as an iframe}
\usage{
as.iframe(widget, file = NULL, selfcontained = TRUE, libdir = NULL,
  width = "100\%", height = 500, `border-width` = "0px", ...)
}
\arguments{
\item{widget}{Widget to save}

\item{file}{File to save HTML into}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}

\item{libdir}{Directory to copy HTML dependencies into (defaults to
  filename_files).}

\item{width}{width of iframe}

\item{height}{height of iframe}

\item{border-width}{width of the borders. default is 0px}

\item{...}{css tag attributes to add to the iframe}
}
\description{
This function allows a widget to be embedded in an Rmd document as an iframe.
This is especially useful when there is potential for conflicts between
assets on the parent page and those used by the widget.
}

