% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDependency}
\alias{getDependency}
\title{Get js and css dependencies for an htmlwidget}
\usage{
getDependency(name, package = name)
}
\arguments{
\item{name}{either an htmlwidget object or the name of a widget.}

\item{package}{name of the package, defaults to the widget name.}
}
\description{
Get js and css dependencies for an htmlwidget
}
