% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JS}
\alias{JS}
\title{Mark elements of a list as a javascript literal.}
\usage{
JS(x)
}
\arguments{
\item{x}{character vector that needs to be treated as a javascript object.}
}
\description{
This function \code{JS()} marks a character vector with a special class, so that
it will be treated as a javascript literal when evaluated on the client-side.
The function \code{JSEvals()} will then identify the elements of a
list that should be evaluated as javascript objects on the client side.
}
\author{
Yihui Xie
}

